requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
psycopg2-binary
pydantic
ollama
python-dotenv
PyMuPDF
pytesseract
pdf2image
Pillow
# Test dependencies
pytest
httpx
pytest-mock
